Moved to the infrastructure middleware namespace.
